+++ src/geometry/rubik_piece.rs
@@ -57,7 +57,7 @@ impl Mesh {
             Vertex::new([-1.0, -1.0, -1.0], [0.0, -1.0, 0.0], back_color),
             Vertex::new([1.0, -1.0, -1.0], [0.0, -1.0, 0.0], back_color),
         ];
-        let index_data: &[u16] = &[
+        let index_data: &[u32] = &[
             0, 1, 2, 2, 3, 0, // top
             4, 5, 6, 6, 7, 4, // bottom
             8, 9, 10, 10, 11, 8, // right
